# collab-service/internal/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.NoteDTO:
    properties:
      body:
        type: string
      folder_id:
        type: string
      title:
        type: string
    required:
    - body
    - folder_id
    - title
    type: object
  dto.ShareDTO:
    properties:
      accessRole:
        $ref: '#/definitions/models.AccessRole'
      userId:
        type: string
    required:
    - accessRole
    - userId
    type: object
  models.AccessRole:
    enum:
    - READ
    - WRITE
    type: string
    x-enum-varnames:
    - AccessLevelRead
    - AccessLevelWrite
  models.Folder:
    properties:
      id:
        type: string
      name:
        type: string
      notes:
        items:
          $ref: '#/definitions/models.Note'
        type: array
      owner:
        $ref: '#/definitions/models.User'
      ownerID:
        type: string
    type: object
  models.Note:
    properties:
      body:
        type: string
      folder:
        $ref: '#/definitions/models.Folder'
      folderID:
        type: string
      id:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      ownerID:
        type: string
      title:
        type: string
    type: object
  models.Roster:
    properties:
      id:
        type: string
      isLeader:
        type: boolean
      team:
        $ref: '#/definitions/models.Team'
      teamID:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: string
    type: object
  models.Team:
    properties:
      id:
        type: string
      rosters:
        items:
          $ref: '#/definitions/models.Roster'
        type: array
      teamName:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      folders:
        items:
          $ref: '#/definitions/models.Folder'
        type: array
      id:
        type: string
      passwordHash:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      rosters:
        items:
          $ref: '#/definitions/models.Roster'
        type: array
      username:
        type: string
    type: object
  models.UserRole:
    enum:
    - MANAGER
    - MEMBER
    type: string
    x-enum-varnames:
    - RoleManager
    - RoleMember
host: localhost:8080
info:
  contact: {}
  description: API for managing teams, folders, notes, and sharing.
  title: Collab Service API
  version: "1.0"
paths:
...
  /folders/{folderId}/share:
    post:
      consumes:
      - application/json
      description: Share a folder with another user with read or write access
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      - description: User ID and permission
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ShareDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties: true
            type: object
      summary: Share folder with user
      tags:
      - folders
  /folders/{folderId}/share/{userId}:
    delete:
      description: Remove a user's access to a shared folder
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with message
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties: true
            type: object
      summary: Revoke shared folder access
      tags:
      - folders
  /notes:
    post:
      consumes:
      - application/json
      description: Create a new note in a folder
      parameters:
      - description: Note data
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/dto.NoteDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created note
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: User not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - notes
  /notes/{noteId}:
    delete:
      parameters:
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete a note
      tags:
      - notes
    get:
      parameters:
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get a note
      tags:
      - notes
    put:
      consumes:
      - application/json
      parameters:
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      - description: Updated note
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/dto.NoteDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
  /notes/{noteId}/share:
    post:
      consumes:
      - application/json
      description: Share a single note with read or write access
      parameters:
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      - description: User ID and permission
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ShareDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with message
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties: true
            type: object
      summary: Share a note with another user
      tags:
      - notes
  /notes/{noteId}/share/{userId}:
    delete:
      description: Remove a user's access to a shared note
      parameters:
      - description: Note ID
        in: path
        name: noteId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with message
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties: true
            type: object
      summary: Revoke note sharing
      tags:
      - notes
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token like: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"